migration to utils.py & graph.py
